+++ src/main/kotlin/racingcar/domain/Car.kt
@@ -1,4 +1,4 @@
-package racingcar.application
+package racingcar.domain
 
 class Car(
     val name: String,
