+++ src/main/kotlin/racingcar/domain/CarMovementDecider.kt
@@ -1,4 +1,4 @@
-package racingcar.application
+package racingcar.domain
 
 fun interface CarMovementDecider {
     fun canMove(): Boolean
