+++ src/test/kotlin/racingcar/domain/RaceBoardTest.kt
@@ -1,8 +1,8 @@
-package racingcar.application
+package racingcar.domain
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
-import racingcar.application.RaceResultPerRound.Companion.fromCars
+import racingcar.domain.RaceResultPerRound.Companion.fromCars
 
 class RaceBoardTest : StringSpec({
     "findWinners() 는 포지션이 가장 높은 차를 반환한다" {
